Make sure that the gateway is in the ARP cache
When imds-filterd launches, it looks up the gateway used for packets
going to the IMDS, and then extracts the gateway's MAC address from
the host ARP cache.  In the event that imds-filterd starts before any
other traffic has passed over the network -- which is possible since
we want it running early in the boot process in order to intercept
traffic from other processes -- this ARP cache entry might not be
present yet.  If this happens, imds-filterd will not be able to send
outgoing packets to the IMDS since it will not have the necessary MAC
address for the ethernet segments it injects.

Sending an ICMP Echo Request packet prompts the kernel to discover
the necessary address and place it into the cache for us.